+++ changelog.md
@@ -1,7 +1,10 @@
 ### CHANGE LOG:
 
-**1.0.0 May 19, 2018:**
-- Create widget with the first release.
+**2.0.0 May 22, 2018:**
+- Delete an ActiveDataProvider. And instead, it simply uses the data parameter (array of ActiveRecord).
 
 **1.1.0 May 20, 2018:**
 - Add sub item and sub container options to separate them with the main container and item options.
+
+**1.0.0 May 19, 2018:**
+- Create widget with the first release.